+++ zcash_client_backend/src/data_api/testing.rs
@@ -26,7 +26,10 @@ use ::sapling::{
 use zcash_address::ZcashAddress;
 use zcash_keys::{
     address::{Address, UnifiedAddress},
-    keys::{UnifiedAddressRequest, UnifiedFullViewingKey, UnifiedSpendingKey},
+    keys::{
+        UnifiedAddressRequest, UnifiedFullViewingKey, UnifiedIncomingViewingKey,
+        UnifiedSpendingKey,
+    },
 };
 use zcash_note_encryption::Domain;
 use zcash_primitives::{
@@ -3347,6 +3350,16 @@ impl WalletWrite for MockWalletDb {
         _unified_key: &UnifiedFullViewingKey,
         _birthday: &AccountBirthday,
         _purpose: AccountPurpose,
+        _key_source: Option<&str>,
+    ) -> Result<Self::Account, Self::Error> {
+        todo!()
+    }
+
+    fn import_account_uivk(
+        &mut self,
+        _account_name: &str,
+        _unified_key: &UnifiedIncomingViewingKey,
+        _birthday: &AccountBirthday,
         _key_source: Option<&str>,
     ) -> Result<Self::Account, Self::Error> {
         todo!()