+++ zcash_client_sqlite/src/wallet/init/migrations/receiving_key_scopes.rs
@@ -549,6 +549,7 @@ mod tests {
             None,
             tx,
             &[(account_id, ufvk0)].into_iter().collect(),
+            &std::collections::HashMap::new(),
         );
 
         db_data